12/requirements.txt
pytest
fastapi[all]
sqlalchemy
testcontainers[postgresql] 
psycopg2-binary
pytest-playwright
